refactor: rebuild the folder structure
