+++ news-api/src/shared/layouts/MainLayout.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react';
-import {Header} from "../../widgets/LayoutHeader/Header.tsx";
-import {Footer} from "../../widgets/ LayoutFooter/Footer.tsx";
-import {PostList} from "../../widgets/PostList/PostList.tsx";
+import {Header} from "../../widgets";
+import {Footer} from "../../widgets";
+import {PostList} from "../../widgets/PostList";
 
 export const MainLayout: FC = () => {
     return (
