+++ news-api/src/widgets/PostList/PostList.tsx
@@ -1,11 +1,11 @@
-import {PostCard} from "../../entities/post/ui/PostCard.tsx";
-import {mockData} from "./mockData.ts";
-import css from './style.module.css'
+import {PostCard} from "../../entities";
+import {MOCK_DATA} from "./lib";
+import css from './post-list.module.css'
 
 export function PostList() {
     return (
         <div className={css.list}>
-        {mockData.map((el) => (
+        {MOCK_DATA.map((el) => (
                 <PostCard key={el.title} postInfo={el} />
             ))}
         </div>
