+++ news-api/src/widgets/PostList/index.ts
@@ -1,0 +1,2 @@
+export {PostList} from './PostList.tsx'
+export {MOCK_DATA} from './lib'
