+++ news-api/src/widgets/PostList/lib/index.ts
@@ -1,0 +1,1 @@
+export { MOCK_DATA} from './MOCK_DATA'
