+++ news-api/src/widgets/index.ts
@@ -1,0 +1,2 @@
+export {Header} from './LayoutHeader'
+export {Footer} from './ LayoutFooter'
